add paths and reorganize folders
