+++ app.js
@@ -15,4 +15,5 @@ app.use('/', (req, res) => {
   res.status(404).json({ message: 'Запрашиваемый ресурс не найден' });
 });
 
+
 app.listen(PORT);
